+++ autorally_control/include/autorally_control/path_integral/gpu_err_chk.h
@@ -1,0 +1,71 @@
+/*
+* Software License Agreement (BSD License)
+* Copyright (c) 2013, Georgia Institute of Technology
+* All rights reserved.
+*
+* Redistribution and use in source and binary forms, with or without
+* modification, are permitted provided that the following conditions are met:
+*
+* 1. Redistributions of source code must retain the above copyright notice, this
+* list of conditions and the following disclaimer.
+* 2. Redistributions in binary form must reproduce the above copyright notice,
+* this list of conditions and the following disclaimer in the documentation
+* and/or other materials provided with the distribution.
+*
+* THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
+* AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+* IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+* DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
+* FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+* DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+* SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+* CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+* OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+* OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+*/
+/**********************************************
+ * @file gpu_err_chk.h
+ * @author Grady Williams <gradyrw@gmail.com>
+ * @date May 24, 2017
+ * @copyright 2017 Georgia Institute of Technology
+ * @brief Generic CUDA error checking wrappers
+ ***********************************************/
+
+#ifndef HANDLE_ERROR_H_
+#define HANDLE_ERROR_H_
+
+#include <cuda_runtime.h>
+
+inline void gpuAssert(cudaError_t code, const char *file, int line, bool abort=true)
+{
+   if (code != cudaSuccess) 
+   {
+      fprintf(stderr,"GPUassert: %s %s %d\n", cudaGetErrorString(code), file, line);
+      if (abort) exit(code);
+   }
+}
+
+inline void __cudaCheckError( const char *file, const int line )
+{
+  cudaError err = cudaGetLastError();
+  if ( cudaSuccess != err )
+  {
+    fprintf( stderr, "cudaCheckError() failed at %s:%i : %s\n",
+              file, line, cudaGetErrorString( err ) );
+    exit( -1 );
+  }
+
+  // More careful checking. However, this will affect performance.
+  // Comment away if needed.
+  err = cudaDeviceSynchronize();
+  if( cudaSuccess != err ) {
+    fprintf( stderr, "cudaCheckError() with sync failed at %s:%i : %s\n",
+    file, line, cudaGetErrorString( err ) );
+    exit( -1 );
+  }
+}
+
+#define CudaCheckError() __cudaCheckError( __FILE__, __LINE__ )
+#define HANDLE_ERROR(ans) { gpuAssert((ans), __FILE__, __LINE__); }
+
+#endif /*HANDLE_ERROR_H_*/